requirements.txt
# UCI Research Intelligence System - Requirements
# Python 3.12 compatible
# Last updated: 2025-01-21

# ==================== AWS & Cloud ====================
boto3==1.34.14
botocore==1.34.14
awscli==1.32.14

# ==================== Data Processing ====================
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.3.2
pyarrow==14.0.2  # For parquet file support

# ==================== Data Generation ====================
faker==22.0.0
faker-edu==1.0.0  # Educational data generation

# ==================== Document Processing ====================
PyPDF2==3.0.1
pdfplumber==0.10.3  # Better PDF extraction
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2
chardet==5.2.0  # Character encoding detection

# ==================== Vector Search & Embeddings ====================
# OpenSearch
opensearch-py==2.4.2
opensearch-dsl==2.1.0

# Embeddings (choose based on your needs)
sentence-transformers==2.2.2
torch==2.2.2  # CPU version, for GPU use torch==2.2.2+cu118
transformers==4.37.2
faiss-cpu==1.8.0  # For local vector search

# Vector Databases
# chromadb==0.4.24  # Commented out due to macOS compilation issues
pinecone-client==3.0.2

# ==================== RAG & LLM Orchestration ====================
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langsmith==0.1.17  # LangChain monitoring

# LLM SDKs
openai==1.8.0
anthropic==0.8.1
cohere==4.44.0  # Alternative LLM option

# ==================== Frontend ====================
streamlit==1.29.0
streamlit-chat==0.1.1  # Chat UI components
streamlit-aggrid==0.3.4  # Advanced tables
plotly==5.18.0
altair==5.2.0

# ==================== Web Backend ====================
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
httpx==0.26.0
aiofiles==23.2.1

# ==================== Database ====================
sqlalchemy==2.0.25
alembic==1.13.1  # Database migrations
psycopg2-binary==2.9.9
redis==5.0.1  # For caching

# ==================== Environment & Config ====================
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# ==================== Authentication & Security ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# ==================== Monitoring & Logging ====================
structlog==24.1.0
python-json-logger==2.0.7
prometheus-client==0.19.0
sentry-sdk==1.39.2  # Error tracking

# ==================== Testing ====================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-env==1.1.3
moto==4.2.12  # Mock AWS services
fakeredis==2.20.1  # Mock Redis
responses==0.24.1  # Mock HTTP requests

# ==================== Development Tools ====================
ipython==8.19.0
jupyter==1.0.0
notebook==7.0.7
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3
pre-commit==3.6.0
autopep8==2.0.4
isort==5.13.2

# ==================== Utilities ====================
click==8.1.7
tqdm==4.66.1
rich==13.7.0  # Beautiful terminal output
requests==2.31.0
urllib3==2.0.7
aiohttp==3.9.1
tenacity==8.2.3  # Retry logic
schedule==1.2.0  # Job scheduling
python-dateutil==2.8.2
pytz==2023.3.post1

# ==================== Data Validation ====================
jsonschema==4.20.0
cerberus==1.3.5
marshmallow==3.20.1

# ==================== Performance ====================
cachetools==5.3.2
joblib==1.3.2  # Parallel processing
asyncio-throttle==1.0.2
aiocache==0.12.2